Move test client to API conftest
